# openapi-vasws.yaml
openapi: 3.0.3
info:
  title: VAS Web Service
  description: >
    VAS Web Service used by Selfcare and CCS to manage subscriber VAS
    registrations. All endpoints are `GET`, URL-based
    (`/vasws/{function}?param=value`), return UTF-8 JSON, and are protected by
    HTTP Basic Authentication. Services are independent: a subscriber may hold
    any combination of them at once, and subscribing to or leaving one never
    affects another. Every subscribe/unsubscribe cycle is retained and returned
    by the history endpoint.
  version: 2.0.0
servers:
  - url: "{baseUrl}"
    description: Base URL of the VAS Web Service (provided by the operator).
    variables:
      baseUrl:
        default: https://example.com
        description: Scheme and host where the service is hosted.
security:
  - BasicAuth: []
paths:
...
  /vasws/removeall:
    get:
      operationId: removeAll
      summary: Remove all of a subscriber's active services
      description: >
        Deactivates the subscriber in every service they are active in, logging
        each to that service's history the same way as `remove`. Succeeds even
        when nothing is active.
      parameters:
        - $ref: '#/components/parameters/Mdn'
      responses:
        '200':
          description: Removal outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationResult'
              example:
                result: 0
                msg: services have been removed successfully
                success: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/SystemError'
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: >
        HTTP Basic credentials (username and password) issued to Selfcare/CCS
        by the VAS Web Service operator.
  parameters:
    Mdn:
      name: mdn
      in: query
      required: true
      description: >
        Subscriber MSISDN. Normalized server-side — digits and an optional
        leading `+` are kept; without a `+`, `+249` (Sudan) is prepended after
        stripping leading zeros.
      schema:
        type: string
      example: "0999900046"
...
  schemas:
    ResultCode:
      type: integer
      description: >
        Shared result code — 0: success, 1: already registered in this service,
        2: not registered in this service, 10: system error.
      enum: [0, 1, 2, 10]
      example: 0
...
    MutationResult:
      type: object
      properties:
        result:
          $ref: '#/components/schemas/ResultCode'
        msg:
          type: string
        success:
          type: boolean
      required: [result, msg, success]
  responses:
    BadRequest:
      description: A required parameter is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MutationResult'
          example:
            result: 10
            msg: mdn is required
            success: false
    Unauthorized:
      description: Missing or incorrect HTTP Basic credentials
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              msg:
                type: string
                example: unauthorized
            required: [success, msg]
          example:
            success: false
            msg: unauthorized
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Basic realm="vasws"
    SystemError:
      description: Unexpected database or system error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MutationResult'
          example:
            result: 10
            msg: system error
            success: false